# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grove Street - Arabic Gaming Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order (purchase)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /marketplace:
    get:
      operationId: listListings
      tags: [marketplace]
      summary: List approved account listings
      parameters:
        - name: gameId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountListing"
    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create account listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListing"
  /marketplace/{id}:
    get:
      operationId: getListing
      tags: [marketplace]
      summary: Get listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListing"
  /marketplace/{id}/approve:
    post:
      operationId: approveListing
      tags: [marketplace]
      summary: Approve listing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListing"
  /marketplace/{id}/reject:
    post:
      operationId: rejectListing
      tags: [marketplace]
      summary: Reject listing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListing"
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, userId, productId, productName, price, status, pointsEarned, createdAt]
      properties:
        id:            { type: integer }
        userId:        { type: integer }
        userName:      { type: string }
        productId:     { type: integer }
        productName:   { type: string }
        productNameAr: { type: string }
        price:         { type: number }
        status:        { type: string }
        userInputData: { type: string }
        pointsEarned:  { type: integer }
        notes:         { type: string }
        createdAt:     { type: string }
    OrderInput:
      type: object
      required: [productId]
      properties:
        productId:     { type: integer }
        userInputData: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        notes:  { type: string }
...
    AccountListing:
      type: object
      required: [id, userId, gameId, title, price, status, createdAt]
      properties:
        id:             { type: integer }
        userId:         { type: integer }
        sellerName:     { type: string }
        sellerVerified: { type: boolean }
        gameId:         { type: integer }
        gameName:       { type: string }
        gameNameAr:     { type: string }
        title:          { type: string }
        description:    { type: string }
        price:          { type: number }
        rank:           { type: string }
        level:          { type: integer }
        region:         { type: string }
        imageUrls:      { type: string }
        status:         { type: string }
        createdAt:      { type: string }
    AccountListingInput:
      type: object
      required: [gameId, title, price]
      properties:
        gameId:      { type: integer }
        title:       { type: string }
        description: { type: string }
        price:       { type: number }
        rank:        { type: string }
        level:       { type: integer }
        region:      { type: string }
        imageUrls:   { type: string }
...